requirements.txt
google-auth>=2.0.0
google-api-python-client>=2.0.0
google-auth-oauthlib>=0.4.0
requests>=2.25.0